@@ -1,21 +1,18 @@
 {
   "homepage": "https://vakotua.github.io/pokedex",
   "name": "pokedex",
-  "version": "0.1.0",
+  "version": "0.2.1",
   "private": true,
   "dependencies": {
-    "-": "^0.0.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "bootstrap": "^5.1.0",
-    "pokedex-promise-v2": "^3.3.2",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0-beta.6",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.0",
-    "save-dev": "^0.0.1-security",
-    "web-vitals": "^1.1.2"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.27.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-icons": "^4.4.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "predeploy": "npm run build",
